+++ djangocms_navigation/views.py
@@ -8,6 +8,8 @@ from django.views.generic import View
 from cms.models import Page
 from cms.utils import get_current_site, get_language_from_request
 
+from djangocms_versioning.constants import ARCHIVED, UNPUBLISHED
+
 from djangocms_navigation.utils import is_model_supported, supported_models
 
 
@@ -32,8 +34,17 @@ class ContentObjectSelect2View(View):
         if not is_model_supported(self.menu_content_model, model):
             return HttpResponseBadRequest()
 
+        queryset_data = self.get_data()
+
+        # Removing unpublished pages from queryset
+        if model == Page:
+            queryset_data = [
+                page for page in queryset_data
+                if not getattr(page.get_title_obj().versions.first(), "state", None) in [ARCHIVED, UNPUBLISHED]
+            ]
+
         data = {
-            "results": [{"text": str(obj), "id": obj.pk} for obj in self.get_data()]
+            "results": [{"text": str(obj), "id": obj.pk} for obj in queryset_data]
         }
         return JsonResponse(data)